fix wrong sampling
also remove a few api and rename spin correlation function
